Normal collision

+++ app/src/main/java/wastedgames/proviant/objects/environment/BuildingUnit.java
@@ -3,7 +3,7 @@ package wastedgames.proviant.objects.environment;
 import wastedgames.proviant.objects.PortableUnit;
 
 public class BuildingUnit extends PortableUnit {
-    public BuildingUnit(int x, int y) {
+    public BuildingUnit(float x, float y) {
         super(x, y);
     }
 }
